feat: Custom opacity for icons
